Revert "[Unit/Tools]: fix issues on  culculation GetHealthPct & GetManaPct"
This reverts commit 7a47ffcce6205e0750352ed4b5113bb232dffef7.

+++ src/server/game/Entities/Unit/Unit.h
@@ -1462,8 +1462,8 @@ class Unit : public WorldObject
         bool HealthBelowPctDamaged(int32 pct, uint32 damage) const { return int64(GetHealth()) - int64(damage) < int64(CountPctFromMaxHealth(pct)); }
         bool HealthAbovePct(int32 pct) const { return GetHealth() > CountPctFromMaxHealth(pct); }
         bool HealthAbovePctHealed(int32 pct, uint32 heal) const { return uint64(GetHealth()) + uint64(heal) > CountPctFromMaxHealth(pct); }
-        float GetHealthPct() const { return GetMaxHealth() ? 100.f * float(GetHealth() / GetMaxHealth()) : 0.0f; }
-        float GetManaPct() const { return GetMaxPower(POWER_MANA) ? 100.f * float(GetPower(POWER_MANA) / GetMaxPower(POWER_MANA)) : 0.0f; }
+        float GetHealthPct() const { return GetMaxHealth() ? 100.f * GetHealth() / GetMaxHealth() : 0.0f; }
+        float GetManaPct() const { return GetMaxPower(POWER_MANA) ? 100.f * GetPower(POWER_MANA) / GetMaxPower(POWER_MANA) : 0.0f; }
         uint32 CountPctFromMaxHealth(int32 pct) const { return CalculatePct(GetMaxHealth(), pct); }
         uint32 CountPctFromCurHealth(int32 pct) const { return CalculatePct(GetHealth(), pct); }
         uint32 CountPctFromMaxMana(int32 pct) const { return CalculatePct(GetMaxPower(POWER_MANA), pct); }
